перенесла бизнеслогику в модели
