+++ src/components/CreateTaskModal/ui/CreateTaskModal.tsx
@@ -1,45 +1,12 @@
-import { postData } from "../../../api";
-import type { TaskResType, TaskUiType } from "../../../types/types.ts";
-import { TaskResSchema, TaskUiSchema } from "../../../types/zodShemas.ts";
-import { useActions } from "../../../store/useActions.ts";
 import { useNavigate } from "react-router";
-import { useTasksState } from "../../../store/slices/tasksSlice.ts";
 import { useParams } from "react-router-dom";
 import closeIcon from "@/assets/icon-cross.svg";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
+import { useCreateTask } from "../model/useCreateTask.ts";
 
 export function CreateTaskModal() {
-  const { addTasks } = useActions();
-  const { tasks } = useTasksState();
   const navigate = useNavigate();
   const { boardId } = useParams() as { boardId: string };
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-    setError,
-  } = useForm<TaskUiType>({
-    resolver: zodResolver(TaskUiSchema),
-  });
-
-  async function onSubmit(data: TaskUiType): Promise<void> {
-    const response: TaskResType | undefined = await postData<
-      TaskResType,
-      TaskUiType
-    >(`boards/${boardId}/tasks/create`, TaskResSchema, data);
-    if (typeof response !== "undefined") {
-      addTasks({
-        board: boardId,
-        tasks: [...tasks[boardId], response],
-      });
-      navigate(`/boards/${boardId}`);
-    } else {
-      setError("status", {
-        message: "Failed to create task. Try again later.",
-      });
-    }
-  }
+  const { register, handleSubmit, errors, onSubmit } = useCreateTask();
 
   return (
     <div className="modal-overlay edit-task-modal">
