+++ src/components/LogInForm/index.ts
@@ -1,0 +1,1 @@
+export { LogInForm } from "./ui/LogInForm.tsx";
