+++ src/components/SignUpForm/index.ts
@@ -1,0 +1,1 @@
+export { SignUpForm } from "./ui/SignUpForm.tsx";
